% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbc.R
\name{cbc}
\alias{cbc}
\title{Calculating CBC}
\usage{
cbc(graph, community, nodename = NULL)
}
\arguments{
\item{igraph}{object}
}
\value{
result
library(SSRM)
}
\description{
Calculating CBC
}
\author{
Yu Hatakeyama
}

