% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-count.R
\name{ds_count}
\alias{ds_count}
\title{Calculating ds_count}
\usage{
ds_count(graph, community, nodename = NULL)
}
\arguments{
\item{igraph}{object}
}
\value{
result
library(SSRM)
}
\description{
Calculating ds_count
}
\author{
Yu Hatakeyama
}

