% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbc.R
\name{lbc}
\alias{lbc}
\title{Calculating LBC}
\usage{
lbc(graph, community, nodename = NULL)
}
\arguments{
\item{igraph}{object}
}
\value{
result
library(SSRM)
}
\description{
Calculating LBC
}
\author{
Yu Hatakeyama
}

