% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leader.R
\name{leader}
\alias{leader}
\title{Calculating leader}
\usage{
leader(graph, visualize = T)
}
\arguments{
\item{igraph}{object}
}
\value{
result
library(SSRM)
}
\description{
Calculating leader
}
\author{
Yu Hatakeyama
}

