% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediacy_score.R
\name{mediacy_score}
\alias{mediacy_score}
\title{Calculating Mediacy Score}
\usage{
mediacy_score(graph, community, nodename = NULL)
}
\arguments{
\item{igraph}{object}
}
\value{
result
library(SSRM)
}
\description{
Calculating Mediacy Score
}
\author{
Yu Hatakeyama
}

