% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediator.R
\name{mediator}
\alias{mediator}
\title{Calculating med_extractor}
\usage{
mediator(mediacy_score_nodes, community)
}
\arguments{
\item{igraph}{object}
}
\value{
result
library(SSRM)
}
\description{
Calculating med_extractor
}
\author{
Yu Hatakeyama
}

