% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbc.R
\name{nbc}
\alias{nbc}
\title{Calculating NBC}
\usage{
nbc(graph, community, nodename = NULL)
}
\arguments{
\item{igraph}{object}
}
\value{
result
library(SSRM)
}
\description{
Calculating NBC
}
\author{
Yu Hatakeyama
}

