% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outermosts.R
\name{outermosts}
\alias{outermosts}
\title{Calculating Outermosts}
\usage{
outermosts(graph, visualize = T)
}
\arguments{
\item{igraph}{object}
}
\value{
result
library(SSRM)
}
\description{
Calculating Outermosts
}
\author{
Yu Hatakeyama
}

