% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outsider.R
\name{outsider}
\alias{outsider}
\title{Calculating Outsider}
\usage{
outsider(graph, community)
}
\arguments{
\item{igraph}{object}
}
\value{
result
library(SSRM)
}
\description{
Calculating Outsider
}
\author{
Yu Hatakeyama
}

