% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssrm-class.R
\docType{data}
\name{SSRM Class}
\alias{SSRM}
\alias{SSRM Class}
\title{SSRM Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SSRM
}
\arguments{
\item{igraph}{object}
}
\value{
result
library(SSRM)
}
\description{
SSRM Class
}
\author{
Yu Hatakeyama
}
\keyword{datasets}

